Fix issues blocking process-pull-request
process-pull-request was failing. Make changes to address the errors.

Information receied from pcolmer:

3 directories have been found with full-stops in name (invalid):

./old_website/docs/2016.04
./old_website/docs/2017.04
./old_website/docs/2016.10
So this has been caused by a security change that Linaro made to all of the
static websites a few weeks ago. As a result of the tighter security, we've
reluctantly had to introduce a rule that means that directory names cannot
have full-stops in them. I thought that I had gone through all of the
repositories for the websites and fixed them but I must have missed this
one. All that needs to be done is to rename the directories to use - or _.

/old_website/mca.html:
   http://www.multicore-association.org/workgroup/oamp.php [404]
This one really is a broken link - that page doesn't exist. Unfortunately,
these things happen when other websites do a clean-up and remove pages,
which is why we have the link checker running on every build, so that dead
links can be removed from our sites.