Adjust URLs for mailman3
Linaro switched all mail list servers to mailman3 in ~2021 Dec.
These use a different URL format than the previous mailman2 deployment.
Adjust URLs in footer and page text for this.

Note: it is unclear exactly which message the govenance page was pointing
to but the one I picked is really the start of the discussion.

Signed-off-by: Bill Mills <bill.mills@linaro.org>