Update charter wording on quorum and voting
To ensure we don't get stuck unable to reach quorum, we discussed at the last OpenAMP Board meeting about updating the charter.  Bill Mills and Nathalie CKC put together proposed wording and the Board unanimously agreed to the proposal.  This commit uploads the new Charter document and updates links pointing to it.